refactor and cleaning import validation step
